Added notes for the changes made from Nokfah's version
This was based on the fixed version of S&D 2.8 provided by Nokfah.
Changes between this version and Nokfah's (available here:
https://goo.gl/a5wp6n):
* When cp mobs are in the same area as you, those mobs are highlighted
in
blue in the cp window
* Added "ma" for mapper area search, and "mf" for mapper world search.
This
functionality already existed but I had a separate alias elsewhere.
* Created a separate query for mob searches that would only include
exact room
name matches, instead of the fuzzy logic used by room name search
* Room searches will display any custom notes you have set for the room
in-line